Questões 6 e 7

+++ InverterLista.hs
@@ -2,6 +2,6 @@ module InverterLista (inverterLista) where
 
 
 -- Função para inverter uma lista
-inverterLista :: [Int] -> [Int]
+inverterLista :: [Double] -> [Double]
 inverterLista [] = []
 inverterLista (x:xs) = inverterLista xs ++ [x]
