+++ ListDoubles.hs
@@ -1,6 +1,6 @@
 module ListDoubles (doubles) where
 
-doubles :: [Int]
+doubles :: [Double]
 doubles = generate 10
     where
         generate x = x : generate (x * 2)
